requirements.txt
fastapi==0.110.0
python-multipart==0.0.9
uvicorn==0.29.0
weaviate-client==4.5.4
python-dotenv==1.0.1
langchain==0.1.16
openai==1.23.6
langchain-openai==0.1.3
pypdf==4.2.0
langchain-community==0.0.34
urllib3<2.0 